Some refactoring so server require and system require match
